Use chat completions for llama.cpp
