+++ pyproject.toml
@@ -13,7 +13,7 @@ dependencies = [
   "rich",
   "r2pipe",
   "inquirer",
-  "llama-cpp-python",
+  "llama-cpp-python==0.3.1",
   "huggingface_hub",
   "appdirs",
   "unidecode",
